+++ src/simmate/workflow_engine/workflow.py
@@ -10,6 +10,7 @@ from pathlib import Path
 import cloudpickle
 import toml
 import yaml
+from django.utils import timezone
 
 import simmate
 from simmate.database.base_data_types import Calculation
@@ -71,10 +72,19 @@ class Workflow:
     `_register_calculation`.
     """
 
-    _parameter_methods = ["run_config", "_run_full"]
+    _parameter_methods: list[str] = ["run_config", "_run_full"]
     """
     List of methods that allow unique input parameters. This helps track where
     `**kwargs` are passed and let's us gather the inputs in one place.
+    """
+
+    exlcude_from_archives: list[str] = []
+    """
+    List of filenames that should be deleted when compressing the output files
+    to a zip file (i.e. when compress_output=True). Any file name is searched
+    for recursively in all subdirectories and removed.
+    
+    For example, VASP calculations remove all POTCAR files from archives.
     """
 
     # -------------------------------------------------------------------------
@@ -151,6 +161,7 @@ class Workflow:
             directory=directory,
             compress_output=compress_output,
             source=source,
+            started_at=timezone.now(),
             **kwargs,
         )
 
@@ -178,13 +189,17 @@ class Workflow:
                 results=results if results != None else {},
                 directory=kwargs_cleaned["directory"],
                 run_id=kwargs_cleaned["run_id"],
+                finished_at=timezone.now(),
             )
 
         # if requested, compresses the directory to a zip file and then removes
         # the directory.
         if compress_output:
             logging.info("Compressing result to a ZIP file.")
-            make_archive(kwargs_cleaned["directory"])
+            make_archive(
+                directory=kwargs_cleaned["directory"],
+                files_to_exclude=cls.exlcude_from_archives,
+            )
 
         # If we made it this far, we successfully completed the workflow run
         logging.info(f"Completed '{cls.name_full}'")
@@ -407,6 +422,7 @@ class Workflow:
         results: dict,
         run_id: str,
         directory: Path,
+        finished_at: timezone.datetime,
     ) -> Calculation:
         """
         Take the output of the `run_config` and any extra information and
@@ -421,6 +437,7 @@ class Workflow:
             run_id=run_id,
             workflow_name=cls.name_full,
             workflow_version=cls.version,
+            finished_at=finished_at,
         )
 
         # Now update the calculation entry with our results. Typically, all of this
@@ -922,6 +939,10 @@ class Workflow:
         register_kwargs_cleaned = cls._deserialize_parameters(
             add_defaults=False, **register_kwargs
         )
+
+        # as an extra, the start time is always registered for ALL calc if given
+        if "started_at" in kwargs.keys():
+            register_kwargs_cleaned["started_at"] = kwargs["started_at"]
 
         # SPECIAL CASE: The exception to the above is with SOURCE, which needs
         # to be in a JSON-serialized form for the database